pyproject.toml
[tool.poetry]
name = "projeto-midgard"
version = "0.1.0"
description = ""
authors = ["FernandoMS4 <fernando381046@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
pyyaml = "^6.0.2"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
